Extract a drawing and calculation of mark image from SnapEditView
into SnapEditMark
